# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API Documentation",
        "description": "API documentation for the EduFlow pedagogical management solution.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EduFlow API Server"
        }
    ],
    "paths": {
...
        "/api/courses/{course}/enroll": {
            "post": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Enroll in a course",
                "operationId": "70df137657a31246717776e7c8b4acc0",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Enrolled Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "enrollment": {
                                            "type": "object"
                                        },
                                        "group": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already Enrolled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/leave": {
            "delete": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Leave a course",
                "operationId": "5d9ac885c675c69a16e249853a710b89",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "You left the course successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enrollment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/enrollments": {
            "get": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Get my enrollments",
                "operationId": "aa55ab3bdcfb63154c783c8d4e6bfd84",
                "responses": {
                    "200": {
                        "description": "List of student enrollments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "enrollments": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/students": {
            "get": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "List students enrolled in a course",
                "operationId": "079b14e283b945fce00637c3819d9ce2",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enrolled students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course_id": {
                                            "type": "integer"
                                        },
                                        "course_title": {
                                            "type": "string"
                                        },
                                        "students": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only the course teacher can view"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "List all groups for a course",
                "operationId": "011ec71f234d5564fed35c4e6bfd5048",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of groups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course_id": {
                                            "type": "integer"
                                        },
                                        "course_title": {
                                            "type": "string"
                                        },
                                        "groups": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only the course teacher can view groups"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/groups/{group}/students": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "List participants in a group",
                "operationId": "4e48fe5c900706c8c4b16e74095d2cc2",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of participants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "group_id": {
                                            "type": "integer"
                                        },
                                        "group_name": {
                                            "type": "string"
                                        },
                                        "course_title": {
                                            "type": "string"
                                        },
                                        "participants": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only the teacher of the parent course can view"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token **_only_**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}